llm-finetune/requirements.txt
# Core deep learning framework
torch>=2.1.0
torchvision>=0.16.0

# Hugging Face ecosystem
transformers>=4.40.0
datasets>=2.18.0
tokenizers>=0.19.0
accelerate>=0.29.0
peft>=0.10.0          # Parameter-Efficient Fine-Tuning (LoRA, QLoRA, etc.)
trl>=0.8.0            # Transformer Reinforcement Learning (SFT, RLHF)

# Quantization
bitsandbytes>=0.43.0  # 4-bit / 8-bit quantization for QLoRA

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.1
evaluate>=0.4.1
sacrebleu>=2.3.3

# Experiment tracking
tensorboard>=2.16.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

# Configuration
pyyaml>=6.0.1
omegaconf>=2.3.0

# Utilities
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.2

# API server (optional inference endpoint)
fastapi>=0.111.0
uvicorn>=0.29.0

# Testing
pytest>=8.1.0
pytest-cov>=5.0.0
